requirements.txt
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3     
attrs==23.1.0
boto3==1.28.67
botocore==1.31.67        
certifi==2023.7.22       
charset-normalizer==3.2.0     
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.11.2
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
frozenlist==1.4.0
idna==3.4
jmespath==1.0.1
multidict==6.0.4
Pillow==10.0.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
typing==3.7.4.3
tzdata==2023.3
urllib3==2.0.4
yarl==1.9.2